plat/arm: scpi: Move to drivers/ folder
