+++ plat/arm/css/drivers/scp/css_pm_scpi.c
@@ -8,10 +8,10 @@
 
 #include <arch_helpers.h>
 #include <common/debug.h>
+#include <drivers/arm/css/css_scpi.h>
 #include <plat/arm/common/plat_arm.h>
 #include <plat/arm/css/common/css_pm.h>
 
-#include "../scpi/css_scpi.h"
 #include "css_scp.h"
 
 /*
